backend/requirements/base.txt
# Web Framework
Django==5.0.0
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Filters and Utils
django-filter==23.5
django-cors-headers==4.3.1

# Date/Time processing
python-dateutil==2.8.2
pytz==2023.3

# Environment and Security
python-decouple==3.8
python-dotenv==1.0.0

# Import/Export
openpyxl==3.1.2
pandas==2.1.4

# Development & Debug
django-debug-toolbar==4.2.0
ipython==8.18.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==21.0.0

# Code Style
black==23.12.0
flake8==6.1.0
isort==5.13.2